// packages-mono/business/components/QuoteTableCell/types.ts
export enum BondQuoteTableColumnKey {
  /** 剩余期限 */
  FirstMaturityDate = 'first_maturity_date',
  /** 代码 */
  BondCode = 'bond_code',
  /** 简称 */
  ShortName = 'short_name',
  /** Bid */
  Bid = 'bid',
  /** Ofr */
  Ofr = 'ofr',
  /** Vol */
  Volume = 'volume',
  /** Broker */
  Broker = 'broker',
  /** 主体评级 */
  IssuerRatingVal = 'issuer_rating_val',
  /** Time */
  UpdateTime = 'update_time',
  /** 债券评级 */
  BondRatingVal = 'bond_rating_val',
  /** 中债净价 */
  ValCleanPrice = 'val_clean_price',
  /** 中债YTM（%） */
  ValYield = 'val_yield',
  /** 中证净价 */
  CsiCleanPrice = 'csi_clean_price',
  /** 中证全价 */
  CsiFullPrice = 'csi_full_price',
  /** 中证YTM（%） */
  CsiYield = 'csi_yield',
  /** CP */
  Cp = 'cp',
  /** 全价 */
  FullPrice = 'full_price',
  /** 净价 */
  CleanPrice = 'clean_price',
  /** 利差 */
  Spread = 'spread',
  /** 含权类型 */
  OptionType = 'option_type',
  /** 上市日 */
  ListedDate = 'listed_date',
  /** 提前还本 */
  RepaymentMethod = 'repayment_method',
  /** 久期 */
  ValModifiedDuration = 'val_modified_duration',
  /** 备注 */
  Comment = 'comment',
  /** 操作人 */
  Operator = 'operator',
  /** 创建日期 */
  CreateTime = 'create_time',
  /** 质押率 */
  ConversionRate = 'conversion_rate',
  /** 中债资信评级 */
  CbcRatingVal = 'cbc_rating_val',
  /** 到期日 */
  MaturityDate = 'maturity_date',
  /** 撤销 */
  ReferType = 'refer_type',
  /** 撤销时间 */
  ReferTime = 'refer_time',
  /** 隐含评级 */
  ImpliedRating = 'implied_rating',
  /** N.Bid */
  NBid = 'n_bid',
  /** Vol.Bid */
  VolumeBid = 'volume_bid',
  /** CP.Bid */
  CpBid = 'cp_bid',
  /** N.Ofr */
  NOfr = 'n_ofr',
  /** Vol.Ofr */
  VolumeOfr = 'volume_ofr',
  /** CP.Bid */
  CpOfr = 'cp_ofr',
  /** Broker(B) */
  BrokerB = 'broker_b',
  /** Broker(O) */
  BrokerO = 'broker_o',
  /** 偏离.Bid */
  OffsetBid = 'offset_bid',
  /** 偏离.Ofr */
  OffsetOfr = 'offset_ofr',
  /** 全价.Bid */
  FullPriceBid = 'full_price_bid',
  /** 净价.Bid */
  CleanPriceBid = 'clean_price_bid',
  /** 利差.Bid */
  SpreadBid = 'spread_bid',
  /** 全价.Ofr */
  FullPriceOfr = 'full_price_ofr',
  /** 净价.Ofr */
  CleanPriceOfr = 'clean_price_ofr',
  /** 利差.Ofr */
  SpreadOfr = 'spread_ofr',
  /** 票面利率 */
  CouponRateCurrent = 'coupon_rate_current',
  /** PVBP */
  ValBasisPointValue = 'val_basis_point_value',
  /** Px */
  Px = 'px',
  /** 交易日 */
  TradedDate = 'traded_date',
  /** 交割日 */
  DeliveryDate = 'delivery_date',
  /** 成交日 */
  DealTime = 'deal_time',
  /** 清算速度 */
  LiquidationSpeed = 'liquidation_speed',
  /** 操作类型 */
  OperationType = 'operation_type',
  /** 操作源 */
  OperationSource = 'operation_source',
  /** 成交状态 */
  DealStatus = 'deal_status',
  /** 发行机构 */
  IssuerInst = 'issuer_inst',
  /** 发行日 */
  IssuerDate = 'issuer_date',
  /** 价格 */
  Price = 'price',
  /** 变动 */
  PriceChange = 'price_change'
}